Make the application of velocity to/from momentum explicitly lmn only
This does not change the program logic, only clarifies that it is
only the LMN solver that works with momentum (i.e. d rho u dt), all
others work with velocity.
The subroutines that convert to/from velocity and momentum exit
early if non-LMN, however the comments in the main program suggest
otherwise.