WIP: refactoring ADIOS2 integration
Essentially replicate the interface concepts in decomp2d&fft, then
we don't need to touch adios2 variables directly, it can all occur
behind the scenes in the 2decomp&fft libraray, therefore user code
is not covered in IFDEF statements

2decomp&fft now has subroutines to:

+ init io - in ADIOS2 this creates an io object, normal builds will
            create the data directory
+ open io - in ADIOS2 create an engine (storing in an internal
            engine registry), normal does nothing
+ start io - in ADIOS2 begins a step, normal does nothing

and the mirror subroutines end and close.
To enable refactoring without doing it all at once have also added
a subroutine to get a pointer to a given engine in the 2decomp&fft
registry, this allows the ADIOS2 version of decomp_2d_write_one to
be used as-is. By doing so have been able to focus on getting this
working for visu.f90 without touching statistcs.f90/tools.f90
(beyond renamings).